# openapi.yaml
openapi: 3.0.3
info:
  title: AlfTekPro Multi-Tenant HRMS API
  version: 1.0.0
  description: |
    Multi-tenant HRMS system with row-level isolation, supporting multiple regions (UAE, USA, India).

    ## Architecture
    - **Authentication**: JWT Bearer tokens with tenant_id in claims
    - **Multi-Tenancy**: Row-level isolation with automatic query filtering
    - **Regions**: UAE (RTL/Arabic), USA (LTR/English), India (LTR/Hindi)

    ## Key Features
    - Employee Management with Temporal Job History
    - Shift & Roster Management
    - Attendance Tracking with Geofencing
    - Leave Management with Approval Workflows
    - Unified Action Center for Approvals
    - Payroll Processing with Regional Logic
    - Asset Management
    - Dynamic Forms (Region-Specific Fields)

  contact:
    name: AlfTekPro Support
    email: support@alftekpro.com
  license:
    name: Proprietary
servers:
  - url: https://api.alftekpro.com/v1
    description: Production Server
  - url: https://dev-api.alftekpro.com/v1
    description: Development Server
  - url: http://localhost:5001/v1
    description: Local Development
...
security:
  - bearerAuth: []
paths:
...
  /attendance/clock-in:
    post:
      tags:
        - Attendance
      summary: Clock in
      description: Record employee clock-in with geofence validation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClockInRequest'
      responses:
        '200':
          description: Clock-in successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      clockInTime:
                        type: string
                        format: time
                      isLate:
                        type: boolean
                      lateByMinutes:
                        type: integer
                  message:
                    type: string
        '400':
          description: Geofence validation failed or already clocked in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /attendance/clock-out:
    post:
      tags:
        - Attendance
      summary: Clock out
      description: Record employee clock-out
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClockOutRequest'
      responses:
        '200':
          description: Clock-out successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      workedHours:
                        type: number
                        format: float
...
  /assets:
    get:
      tags:
        - Assets
      summary: Get all assets
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [Available, Assigned, UnderMaintenance, Disposed]
        - name: assetType
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of assets
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Asset'

    post:
      tags:
        - Assets
      summary: Create asset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAssetRequest'
      responses:
        '201':
          description: Asset created
  /assets/{id}/assign:
    post:
      tags:
        - Assets
      summary: Assign asset to employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignAssetRequest'
      responses:
        '200':
          description: Asset assigned successfully
  /assets/{id}/return:
    post:
      tags:
        - Assets
      summary: Return asset from employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReturnAssetRequest'
      responses:
        '200':
          description: Asset returned successfully
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token with claims: tenant_id, user_id, role

        Example header: Authorization: Bearer <token>
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        errorCode:
          type: string
          description: i18n error code
        errors:
          type: object
          additionalProperties:
            type: string
          description: Validation errors keyed by field name

    # ==================== AUTH SCHEMAS ====================
...
    ClockInRequest:
      type: object
      required:
        - employeeId
        - latitude
        - longitude
      properties:
        employeeId:
          type: string
          format: uuid
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        ipAddress:
          type: string
          description: Auto-captured from request
    ClockOutRequest:
      type: object
      required:
        - employeeId
      properties:
        employeeId:
          type: string
          format: uuid
        ipAddress:
          type: string
...
    Asset:
      type: object
      properties:
        id:
          type: string
          format: uuid
        assetCode:
          type: string
        assetType:
          type: string
        make:
          type: string
        model:
          type: string
        serialNumber:
          type: string
        purchaseDate:
          type: string
          format: date
        purchasePrice:
          type: number
          format: float
        status:
          type: string
          enum: [Available, Assigned, UnderMaintenance, Disposed]
    CreateAssetRequest:
      type: object
      required:
        - assetCode
        - assetType
      properties:
        assetCode:
          type: string
        assetType:
          type: string
        make:
          type: string
        model:
          type: string
        serialNumber:
          type: string
        purchaseDate:
          type: string
          format: date
        purchasePrice:
          type: number
          format: float
    AssignAssetRequest:
      type: object
      required:
        - employeeId
        - assignedDate
        - assignedCondition
      properties:
        employeeId:
          type: string
          format: uuid
        assignedDate:
          type: string
          format: date
        assignedCondition:
          type: string
    ReturnAssetRequest:
      type: object
      required:
        - returnedDate
        - returnedCondition
      properties:
        returnedDate:
          type: string
          format: date
        returnedCondition:
          type: string
        returnNotes:
          type: string
...